Refactor: libcrmcommon: move PID functions to own file
Line counts before:
 1515 lib/common/utils.c

and after:
 1353 lib/common/utils.c
  183 lib/common/pid.c
